+++ src/resources.rs
@@ -2,8 +2,5 @@
 // file (`.rc`) with icon and menu definitions.
 // `main.rs` uses these constants to load the resources during runttime.
 
-pub const ICON_KEYBOARD: u16 = 1;
-pub const ICON_KEYBOARD_DELETE: u16 = 2;
-pub const MENU: u16 = 10;
-pub const MENU_EXIT: u16 = 11;
-pub const MENU_DISABLE: u16 = 12;
+pub const ICON_KEYBOARD: usize = 1;
+pub const ICON_KEYBOARD_DELETE: usize = 2;